requirements.txt
alembic==0.7.6
click==4.0
-e git+https://github.com/mitsuhiko/flask.git@87222087b333c4f9ccc665e787fdf304cabeba73#egg=Flask-master
Flask-Testing==0.4.2
gevent==1.0.2
greenlet==0.4.7
honcho==0.6.6
inflection==0.3.1
ipython==3.1.0
itsdangerous==0.24
Jinja2==2.7.3
Mako==1.0.1
MarkupSafe==0.23
nose==1.3.7
## !! Could not determine repository location
#phlaskr==0.0.1
py-bcrypt==0.4
PyExecJS==1.1.0
SQLAlchemy==1.0.5
Werkzeug==0.10.4
py==1.4.28
pytest==2.7.1
